# jwt: Simplify JWT parsing