@minio-ops 2020-01-13
## Mint Automation
| Test | Result |
|:--- |:--- |
| mint-xl.sh | :heavy_check_mark: |
| mint-fs.sh | :heavy_check_mark: |
| mint-dist-xl.sh | :heavy_check_mark: |
| mint-gateway-nas.sh | :heavy_check_mark: |
| mint-large-bucket.sh | :x: [more...](#ad2d9a4b-9406-4857-aea4-6a2a3a84e13c) |
| mint-worm.sh | :x: [more...](#6d721ef8-1cff-4c56-9b5d-90dbdb5aa377) |
<a name="6d721ef8-1cff-4c56-9b5d-90dbdb5aa377"></a>
### 8802-2f72808/mint-worm.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.55:31324
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            worm
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 8ad7a55b20bd:/mint/log /tmp/mint-logs'

(1/1) Running worm tests ... FAILED in 2 minutes and 1 seconds
{
  "alert": "",
  "args": {
    "bucketName": "bucket-dabhh44382zdsei53sc13439hjsdvkpv130tw37amoxfswp2xrli236",
    "expiry": 60000000000,
    "objectName": "testObject"
  },
  "duration": 120332,
  "error": "RequestError: send request failed\ncaused by: Put http://72.28.97.55:31324/bucket-dabhh44382zdsei53sc13439hjsdvkpv130tw37amoxfswp2xrli236: dial tcp 72.28.97.55:31324: i/o timeout",
  "function": "PutAndDelete",
  "message": "WORM_MODE ON  - CreateBucket Failed",
  "name": "test worm mode",
  "status": "FAIL"
}

Executed 0 out of 1 tests successfully.
```
<a name="ad2d9a4b-9406-4857-aea4-6a2a3a84e13c"></a>
### 8802-2f72808/mint-large-bucket.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.61:31989
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp e2e0e85183a7:/mint/log /tmp/mint-logs'

(1/13) Running awscli tests ... done in 2 minutes and 17 seconds
(2/13) Running aws-sdk-java tests ... done in 1 seconds
(3/13) Running aws-sdk-ruby tests ... done in 7 seconds
(4/13) Running mc tests ... done in 1 minutes and 4 seconds
(5/13) Running minio-go tests ... FAILED in 3 minutes and 47 seconds
{
  "args": {},
  "duration": 112915,
  "error": "Please reduce your request",
  "function": "testStorageClassMetadataCopyObject()",
  "message": "CopyObject failed on RRS",
  "name": "minio-go: testStorageClassMetadataCopyObject",
  "status": "FAIL"
}

Executed 4 out of 13 tests successfully.
```
Deleting image on docker hub
Deleting image locally