@harshavardhana 2020-01-14
@klauspost after some effort :-) 

```
~ benchcmp old.txt new.txt 
benchmark                     old ns/op     new ns/op     delta
BenchmarkParseJWTClaims-4     14734         3769          -74.42%

benchmark                     old allocs     new allocs     delta
BenchmarkParseJWTClaims-4     68             42             -38.24%

benchmark                     old bytes     new bytes     delta
BenchmarkParseJWTClaims-4     4297          2216          -48.43%
```