@harshavardhana 2020-01-14
Okay some more @klauspost I am actually surprised to see hmac using a lot of memory.

```
~ benchcmp old.txt new.txt 
benchmark                     old ns/op     new ns/op     delta
BenchmarkParseJWTClaims-4     6310          3199          -49.30%

benchmark                     old allocs     new allocs     delta
BenchmarkParseJWTClaims-4     68             32             -52.94%

benchmark                     old bytes     new bytes     delta
BenchmarkParseJWTClaims-4     4296          1889          -56.03%
```

```
~ benchstat old.txt new.txt 
name              old time/op    new time/op    delta
ParseJWTClaims-4    6.31µs ± 0%    3.20µs ± 0%   ~     (p=1.000 n=1+1)

name              old alloc/op   new alloc/op   delta
ParseJWTClaims-4    4.30kB ± 0%    1.89kB ± 0%   ~     (p=1.000 n=1+1)

name              old allocs/op  new allocs/op  delta
ParseJWTClaims-4      68.0 ± 0%      32.0 ± 0%   ~     (p=1.000 n=1+1)
```